refactor: split lib.rs into config.rs + module.rs modules
