+++ src/module.rs
@@ -1,0 +1,65 @@
+use napi::bindgen_prelude::{Result, Uint8Array};
+use napi_derive::napi;
+use walrus::{Module, RawCustomSection};
+
+use crate::ModuleConfig;
+
+#[napi]
+pub struct WasmModule {
+  pub(crate) inner: Module,
+}
+
+#[napi]
+impl WasmModule {
+  #[napi(factory)]
+  /// Construct a new module from the given path with the default
+  /// configuration.
+  pub fn from_path(path: String) -> Result<Self> {
+    Ok(Self {
+      inner: Module::from_file(path)?,
+    })
+  }
+
+  #[napi(factory)]
+  /// Construct a new module from the given path and configuration.
+  pub fn from_file_with_config(path: String, config: &ModuleConfig) -> Result<Self> {
+    Ok(Self {
+      inner: Module::from_file_with_config(path, &config.inner)?,
+    })
+  }
+
+  #[napi]
+  /// Emit this module into an in-memory wasm buffer.
+  pub fn emit_wasm(&mut self, demangle: bool) -> Result<Uint8Array> {
+    if demangle {
+      demangle_module(&mut self.inner);
+    }
+    // https://github.com/WebAssembly/tool-conventions/blob/9b80cd2339c648822bb845a083d9ffa6e20fb1ee/BuildId.md
+    if self
+      .inner
+      .customs
+      .iter()
+      .all(|(_, section)| section.name() != "build_id")
+    {
+      self.inner.customs.add(RawCustomSection {
+        name: "build_id".to_string(),
+        data: uuid::Uuid::new_v4().as_bytes().into(),
+      });
+    }
+
+    Ok(self.inner.emit_wasm().into())
+  }
+}
+
+fn demangle_module(module: &mut Module) -> &mut Module {
+  for func in module.funcs.iter_mut() {
+    let name = match &func.name {
+      Some(name) => name,
+      None => continue,
+    };
+    if let Ok(sym) = rustc_demangle::try_demangle(name) {
+      func.name = Some(sym.to_string());
+    }
+  }
+  module
+}
